requirements.txt
Flask
requests
pymongo
python-dotenv
selenium
webdriver-manager
